Type Hardening, Consolidation, Assert Replacement
General:
chore: some code formatting
feat: changed guid definition to hoist the required type conversion, preventing the internal warning generation
feat: replaced macro std assert and macro assert usage with ::hg::assertd/rt
feat: fixed some template types by using underlying container type queries
feat: enforced typed constraints for type erase supported wrappers, preventing instantiation with incomplete types
feat: added explicit warning suppression to reduce warning noise when dealing with highly- and safely constrained type erase objects
feat: added overload struct helper to support type overloaded lambda expressions
feat: changed AutoArray<T, ...> to a concrete template type and added some utility function

+++ src/Editor.Render/Renderer/EditorRenderer.cpp
@@ -21,7 +21,7 @@ EditorRenderer::EditorRenderer(
 	mref<nmpt<engine::gfx::memory::GlobalPooledAllocator>> globalGfxAllocator_
 ) noexcept :
 	Renderer(
-		Guid { "__Static__Proxy"_typeId.data, 0u, 0u, "Editor-Renderer"_typeId.data },
+		Guid { static_cast<u32>("__Static__Proxy"_typeId.data), 0u, 0u, "Editor-Renderer"_typeId.data },
 		"Editor-Renderer",
 		makeCompileGraph(),
 		nullptr,
@@ -56,7 +56,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	beginPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<DummyProvider>>();
+				->allocate<graph::SubPassNode<DummyProvider>>();
 			return node;
 		}
 	);
@@ -64,7 +64,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	uiPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<engine::reflow::render::ReflowPass>>();
+				->allocate<graph::SubPassNode<engine::reflow::render::ReflowPass>>();
 			return node;
 		}
 	);
@@ -72,7 +72,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	endPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<TmpEndPass>>(vk::ImageLayout::ePresentSrcKHR);
+				->allocate<graph::SubPassNode<TmpEndPass>>(vk::ImageLayout::ePresentSrcKHR);
 			return node;
 		}
 	);
